added remaining files to git

+++ app.py
@@ -19,24 +19,25 @@ app.config['SECRET_KEY'] = os.environ.get('SECRET_KEY')
 
 mongo = PyMongo(app)
 
-DEBUG = 'DEVELOPMENT' in os.environ
 
 @app.route('/')
 @app.route('/index')
 def index():
-    recipes = mongo.db.recipes.find()
-
     return render_template('index.html',  title="World Community Cookbook")
-
 
 
 @app.route('/logout')
 def logout():
+    # clear the session and flash a confirmation information message
     session.clear()
     flash('You have logged out successfully - see you soon', 'info')
     return render_template('index.html', title='Logged out')
-    
-''' start of code from Anthony Herbert -> Pretty Printed '''
+
+
+''' The code here '''
+''' is by Anthony Herbert -> Pretty Printed '''
+''' https://www.youtube.com/watch?v=vVx1737auSE '''
+
 
 @app.route('/login', methods=['POST', 'GET'])
 def login():
@@ -51,7 +52,7 @@ def login():
             if bcrypt.hashpw(request.form['pass'].encode('utf-8'), login_user['password']) == login_user['password']:
                 session['username'] = request.form['username']
                 flash(session['username'] + ' you have logged in successfully!', 'success')
-                return render_template("index.html", username=session['username'],
+                return render_template("show_own_recipes.html", login_user=login_user,
                                        recipes=mongo.db.recipes.find(),
                                        courses=mongo.db.meal_courses.find(),
                                        title='Welcome '+session['username'])
@@ -80,9 +81,9 @@ def register():
 
     return render_template('register.html', title='Sign Up!')
 
-
-''' end of code from Anthony Herbert -> Pretty Printed '''
-
+''' End of code '''
+''' by Anthony Herbert -> Pretty Printed '''
+''' https://www.youtube.com/watch?v=vVx1737auSE '''
 
 ''' Recipe section '''
 
@@ -104,6 +105,7 @@ def enter_recipe():
     else:
         flash('You must be logged in to add a recipe', 'warning')
         return render_template('login.html', title='Login to add recipe')
+
 
 @app.route('/insert_recipe', methods=['POST'])
 def insert_recipe():
@@ -136,13 +138,18 @@ def insert_recipe():
 @app.route('/show_one_recipe/<recipe_id>')
 def show_one_recipe(recipe_id):
     the_recipe = mongo.db.recipes.find_one({"_id": ObjectId(recipe_id)})
-    cook_name = mongo.db.recipes.find_one({"_id": ObjectId(recipe_id)})['cook_name']
-    current_user = session['username']
-    user_role = mongo.db.users.find_one({"username": current_user})['role']
-    return render_template("show_one_recipe.html", recipes=the_recipe)
-
-
-
+    if 'username' not in session:
+        current_user = ""
+        return render_template("show_one_recipe.html", recipes=the_recipe, current_user=current_user)
+    else:
+        cook_name = mongo.db.recipes.find_one({"_id": ObjectId(recipe_id)})['cook_name']
+        current_user = session['username']
+        user_role = mongo.db.users.find_one({"username": current_user})['role']
+        return render_template("show_one_recipe.html",
+                               recipes=the_recipe,
+                               current_user=current_user,
+                               user_role=user_role,
+                               cook_name=cook_name)
 
 
 
@@ -150,60 +157,65 @@ def show_one_recipe(recipe_id):
 @app.route('/edit_recipe/<recipe_id>')
 def edit_recipe(recipe_id):
     cuisines = mongo.db.cuisines.find().collation({'locale': 'en'}).sort('cuisine_name', 1)
-    meal_course = mongo.db.meal_course.find().collation({'locale': 'en'}).sort('course_name',1)
+    meal_course = mongo.db.meal_course.find().collation({'locale': 'en'}).sort('course_name', 1)
     main_ingredient = mongo.db.main_ingredient.find().collation({'locale': 'en'}).sort('main_ingredient', 1)
-    recipes=mongo.db.recipes.find()
+    recipes = mongo.db.recipes.find()
     recipe = mongo.db.recipes.find_one({"_id": ObjectId(recipe_id)})
     cook_name = mongo.db.recipes.find_one({"_id": ObjectId(recipe_id)})['cook_name']
-    current_user = session['username']
-
-    user_role = mongo.db.users.find_one({"username": current_user})['role']
-
-
-    if user_role == 'admin':
-        return render_template('edit_recipe.html',
-                            recipe=mongo.db.recipes.find_one({'_id': ObjectId(recipe_id)}),
-                            cuisines=cuisines,
-                            meal_course=meal_course,
-                            main_ingredient=main_ingredient)
-    if current_user == cook_name:
-        return render_template('edit_recipe.html',
-                            recipe=mongo.db.recipes.find_one({'_id': ObjectId(recipe_id)}),
-                            cuisines=cuisines,
-                            meal_course=meal_course,
-                            main_ingredient=main_ingredient)
+    if 'username' not in session:
+        current_user = ""
+        role = ""
+        flash('You must be logged in to edit your own recipes', 'warning')
+        return redirect(url_for('show_recipes', recipes=recipes, role=role))
     else:
-        flash('You can only edit recipes which you have created', 'warning')
-        return render_template('show_recipes.html', recipes=recipes)
-
+        current_user = session['username']
+        user_role = mongo.db.users.find_one({"username": current_user})['role']
+        if user_role == 'admin':
+            return render_template('edit_recipe.html',
+                                   recipe=mongo.db.recipes.find_one({'_id': ObjectId(recipe_id)}),
+                                   cuisines=cuisines,
+                                   meal_course=meal_course,
+                                   main_ingredient=main_ingredient)
+        if current_user == cook_name:
+            return render_template('edit_recipe.html',
+                                   recipe=mongo.db.recipes.find_one({'_id': ObjectId(recipe_id)}),
+                                   cuisines=cuisines,
+                                   meal_course=meal_course,
+                                   main_ingredient=main_ingredient)
+        else:
+            flash(
+                'You can only edit recipes which you have created', 'warning'
+                )
+            return render_template('show_recipes.html', recipe=recipe)
 
 
 @app.route('/update_recipe/<recipe_id>', methods=["POST"])
 def update_recipe(recipe_id):
     recipes = mongo.db.recipes
     req = request.form
-    recipes.update_one({'_id': ObjectId(recipe_id)},
+    recipes.update_one(
+        {'_id': ObjectId(recipe_id)},
         {
-        '$set': {
-         'recipe_name': req.get('recipe_name'),
-         'cook_name': session['username'],
-         'cuisine_name': req.get('cuisine_name'),
-         'main_ingredient': req.get('main_ingredient'),
-         'prep_time': req.get('prep_time'),
-         'total_cooking_time': req.get('total_cooking_time'),
-         'servings': req.get('servings'),
-         'ingredients': req.get('ingredients').splitlines(),
-         'instructions': req.get('instructions').splitlines(),
-         'image_src': req.get('image_src'),
-         'meal_course': req.get('meal_course'),
-         'protein': req.get('protein'),
-         'fat_per_serve': req.get('fat_per_serve'),
-         'carbohydrate': req.get('carbohydrate'),
-         'dietary_fibre': req.get('dietary_fibre'),
-         'cholesterol': req.get('cholesterol'),
-         'energy': req.get('energy')
-        }
-    })
+         '$set': {
+          'recipe_name': req.get('recipe_name'),
+          'cook_name': session['username'],
+          'cuisine_name': req.get('cuisine_name'),
+          'main_ingredient': req.get('main_ingredient'),
+          'prep_time': req.get('prep_time'),
+          'total_cooking_time': req.get('total_cooking_time'),
+          'servings': req.get('servings'),
+          'ingredients': req.get('ingredients').splitlines(),
+          'instructions': req.get('instructions').splitlines(),
+          'image_src': req.get('image_src'),
+          'meal_course': req.get('meal_course'),
+          'protein': req.get('protein'),
+          'fat_per_serve': req.get('fat_per_serve'),
+          'carbohydrate': req.get('carbohydrate'),
+          'dietary_fibre': req.get('dietary_fibre'),
+          'cholesterol': req.get('cholesterol'),
+          'energy': req.get('energy')
+         }
+          })
     recipe = req.get('recipe_name')
     flash(f'Recipe {recipe} updated.', 'success')
     return redirect(url_for('show_recipes'))
@@ -216,12 +228,11 @@ def show_recipes():
                            title='Browse recipes')
 
 
-
-
-
-
-
-
+@app.route('/show_own_recipes')
+def show_own_recipes():
+    recipes = mongo.db.recipes.find()
+    current_user = session['username']
+    return render_template('show_own_recipes.html', current_user=current_user, recipes=recipes)
 
 @app.route('/search_recipes')
 def search_recipes():
@@ -235,18 +246,34 @@ def search():
     return render_template("search_results.html", recipes=recipes)
 
 
-
-
-
-
 @app.route('/delete_recipe/<recipe_id>')
 def delete_recipe(recipe_id):
-    mongo.db.recipes.delete_one({'_id': ObjectId(recipe_id)})
-    flash('Recipe deleted.', 'success')
-    return redirect(url_for('show_recipes'))
-
-
-
+    recipes = mongo.db.recipes.find()
+    recipe = mongo.db.recipes.find_one({"_id": ObjectId(recipe_id)})
+    cook_name = mongo.db.recipes.find_one(
+                                         {"_id": ObjectId(recipe_id)}
+                                         )['cook_name']
+    current_user = session['username']
+    user_role = mongo.db.users.find_one({"username": current_user})['role']
+    if 'username' not in session:
+        current_user = ""
+        role = ""
+        flash('You must be logged in to delete your own recipes', 'danger')
+        return redirect(url_for('show_recipes', recipes=recipes, role=role))
+    else:
+        if user_role == 'admin':
+            mongo.db.recipes.delete_one({'_id': ObjectId(recipe_id)})
+            flash('Recipe deleted.', 'success')
+            return redirect(url_for('show_recipes'))
+        if current_user == cook_name:
+            mongo.db.recipes.delete_one({'_id': ObjectId(recipe_id)})
+            flash('Recipe deleted.', 'success')
+            return redirect(url_for('show_recipes'))
+        else:
+            flash(
+                'You can only delete recipes which you have created', 'danger'
+                )
+            return render_template('show_recipes.html')
 
 
 ''' The routes below handle administration of cuisine types '''
@@ -256,7 +283,8 @@ def delete_recipe(recipe_id):
 def cuisine_admin():
     if 'username' in session:
         return render_template("cuisine_admin.html", title='Cuisines',
-                               cuisines=mongo.db.cuisines.find().collation({'locale': 'en'}).sort('cuisine_name', 1))
+                               cuisines=mongo.db.cuisines.find().collation(
+                                   {'locale': 'en'}).sort('cuisine_name', 1))
     else:
         flash('You must be logged in to add or edit a cuisine', 'warning')
         return render_template('login.html')
@@ -271,13 +299,15 @@ def edit_cuisine(cuisines_id):
     created_by = created_by
     if user_role == 'admin':
         return render_template('edit_cuisine.html',
-                               cuisines=mongo.db.cuisines.find_one({'_id': ObjectId(cuisines_id)}))
+                               cuisines=mongo.db.cuisines.find_one(
+                                   {'_id': ObjectId(cuisines_id)}))
     if created_by != current_user:
         flash('You can only edit cuisines which you have created', 'warning')
         return redirect(url_for('cuisine_admin'))
     else:
         return render_template('edit_cuisine.html',
-                               cuisines=mongo.db.cuisines.find_one({'_id': ObjectId(cuisines_id)}))
+                               cuisines=mongo.db.cuisines.find_one(
+                                   {'_id': ObjectId(cuisines_id)}))
 
 
 @app.route('/update_cuisine/<cuisines_id>', methods=['POST'])
@@ -293,11 +323,12 @@ def update_cuisine(cuisines_id):
                 return redirect(url_for('cuisines', cuisines=cuisines))
         else:
             cuisine.update_one({'_id': ObjectId(cuisines_id)},
-                    {
-                    '$set': {
-                'cuisine_name': request.form.get('cuisine_name'),
-                'created_by': session['username']
-                }})
+                               {
+                                '$set': {
+                                 'cuisine_name': request.form.get(
+                                     'cuisine_name'),
+                                 'created_by': session['username']
+                                }})
             flash('Your cuisine has been updated')
 
             return redirect(url_for('cuisines'))
@@ -308,7 +339,8 @@ def update_cuisine(cuisines_id):
 @app.route('/delete_cuisine/<cuisines_id>', methods=['POST', 'GET'])
 def delete_cuisine(cuisines_id):
     cuisines = mongo.db.cuisines
-    created_by = cuisines.find_one({"_id": ObjectId(cuisines_id)})['created_by']
+    created_by = cuisines.find_one(
+        {"_id": ObjectId(cuisines_id)})['created_by']
     current_user = session['username']
     user_role = mongo.db.users.find_one({"username": current_user})['role']
     created_by = created_by
@@ -318,7 +350,6 @@ def delete_cuisine(cuisines_id):
         return redirect(url_for('cuisine_admin'))
     if created_by != current_user:
         flash('You can only delete cuisines which you have created', 'warning')
-
         return redirect(url_for('cuisine_admin'))
     else:
         cuisines.delete_one({'_id': ObjectId(cuisines_id)})
@@ -348,8 +379,12 @@ def insert_cuisine():
 
 @app.route('/add_cuisine')
 def add_cuisine():
-    return render_template('add_cuisine.html', title='Add New Cuisine')
+    if 'username' not in session:
+        return render_template('index.html', title="World Community Cookbook")
+    else:
+        return render_template('add_cuisine.html', title='Add New Cuisine')
 
 
 if __name__ == "__main__":
-    app.run(host=os.environ.get("IP"), port=os.environ.get("PORT"), debug=True)
+    app.run(host=os.environ.get("IP"),
+    port=os.environ.get("PORT"), debug=True)
